Improve handling no tags via postgres defaults
Previously, we'd create an array and return it. If we appended to this
array, it wasn't saved to the attribute. We could first set the value
to an empty array, but then it would have to account for detecting
`changed?` and similar. Alternatively, using postgres defaults to
control the behavior is a cleaner approach that doesn't result in
unexpected behavior.